requirements.txt
plaid-python
requests